logic set favourite team logic created
